bring dashboards and reports menus one step above
