+++ openbb_terminal/__init__.py
@@ -6,6 +6,6 @@ from . import (
     etf,
     forex,
     portfolio,
-    jupyter,
+    reports,
     stocks,
 )
