+++ openbb_terminal/dashboards/widgets/currencies.json
