+++ openbb_terminal/reports/OpenBB_reports_logo.png
