+++ openbb_terminal/reports/widget_helpers.py
