+++ openbb_terminal/reports/widgets/report.css
