+++ openbb_terminal/reports/widgets/report.j2
