+++ openbb_terminal/reports/widgets/style.css
